\name{R330-package}
\alias{R330-package}
\alias{R330}
\docType{package}
\title{
Useful functions for Stats 330
}
\description{
Contains useful functions and data for the Stats 330 course at the University of Auckland
}
\details{
\tabular{ll}{
Package: \tab R330\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-01-13\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Alan Lee, Blair Robertson

Maintainer: Alan Lee <aj.lee@auckland.ac.nz >
}

\keyword{ package }
